Split up templateEditing.test.ts
It was just getting too big to be a single file.
